// src/app/models/sjcs/ComisariaItem.ts
export class ComisariaItem {

    idComisaria: string;
    nombre: string;
    idInstitucion: string;
    domicilio: string;
    fechabaja: string;
    codigoPostal: string;
    idProvincia: string;
    nombreProvincia: string;
    idPoblacion: string;
    nombrePoblacion: string;
    telefono1: string;
    telefono2: string;
    fax1: string;
    email: string;
    codigoExt: string;
    historico: boolean;
    usuModificacion: string;
    fechaModificacion: string;
    visibleMovil;

    constructor() { }
}